ETL/requirements.txt
pandas==2.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy>=1.4